requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
